Mejorando retroalimentacion

+++ src/app/components/one-post.jsx
@@ -5,39 +5,43 @@ class OnePost extends React.Component{
         super(props);
 
         this.deletePost = this.deletePost.bind(this);
-    }
+    }    
 
 
     deletePost(e){
-        fetch(`/api/post/${this.props.id}`,{
-            method: 'DELETE',
-            //body: JSON.stringify( this.key ),
-            headers: {
-                'Accept' : 'application/json',
-                'Content-type' : 'application/json'
-            }
-        }).then(
-            res => res.json()
-        ).then(
-            data => {
-                if( data.error ){
-                    //this.setState({error: data.msg });
-                }else{                    
-                    this.props.fetchPosts();                    
+        if( confirm('Esta seguro de querer eliminar este post') ){
+            fetch(`/api/post/${this.props.id}`,{
+                method: 'DELETE',
+                //body: JSON.stringify( this.key ),
+                headers: {
+                    'Accept' : 'application/json',
+                    'Content-type' : 'application/json'
                 }
-            }
-        ).catch(
-            err => console.log( err )
-        );
+            }).then(
+                res => res.json()
+            ).then(
+                data => {
+                    if( data.error ){
+                        //this.setState({error: data.msg });
+                    }else{                   
+                        M.toast({ html: 'Post Eliminado', classes : 'success' });
+                        this.props.fetchPosts();                    
+                    }
+                }
+            ).catch(
+                err => console.log( err )
+            );
+        }
+        
         e.preventDefault();
     }
 
     render(){
         let date = new Date(this.props.date_create);
         
-        return(
+        return(                        
             <div className="user-item">
-                <div className="head-item">
+                <div className="head-item">                    
                     <div className="item-id">{this.props.author} ({this.props.mail})</div>
                     <div className="item-date-create">{date.toDateString()}</div>
                 </div>
@@ -46,8 +50,8 @@ class OnePost extends React.Component{
                     <form method="post" onSubmit={this.deletePost} >
                         <input type="submit" className="btn canvas delete-me sm " value="eliminar"/>
                     </form>
-                </div>                                
-            </div>
+                </div> 
+            </div>            
         )
     };
 }
